+++ src/main/java/nora/movlog/controller/HomeController.java
@@ -1,17 +1,31 @@
 package nora.movlog.controller;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import nora.movlog.service.user.MemberService;
+import nora.movlog.utils.dto.user.MemberJoinRequestDto;
+import nora.movlog.utils.dto.user.MemberLoginRequestDto;
+import nora.movlog.utils.validators.MemberValidator;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @Slf4j
 @RequiredArgsConstructor
 @Controller
 public class HomeController {
 
+    private final MemberService memberService;
+    private final MemberValidator memberValidator;
+
+
+
+    // 홈
     @GetMapping(value={"", "/"})
     public String home(Model model, Authentication auth) {
 
@@ -19,5 +33,30 @@ public class HomeController {
     }
 
 
+    // 회원가입
+    @GetMapping("/join")
+    public String joinPage(Model model) {
+        model.addAttribute("memberJoinRequest", new MemberJoinRequestDto());
 
+        return "joinPage";
+    }
+
+    @PostMapping("/join")
+    public String joinPage(@Valid @ModelAttribute MemberJoinRequestDto dto,
+                           BindingResult bindingResult) {
+        if (memberValidator.validateJoin(dto, bindingResult).hasErrors())
+            return "redirect:/join";
+
+        memberService.join(dto);
+
+        return "redirect:/login";
+    }
+
+    // 로그인 페이지
+    @GetMapping("/login")
+    public String loginPage(Model model) {
+        model.addAttribute("memberLoginRequest", new MemberLoginRequestDto());
+
+        return "loginPage";
+    }
 }
